% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcimp.R
\docType{data}
\name{NONE}
\alias{NONE}
\alias{PIPE}
\alias{BUFFLE}
\title{Radiation Impedance Calculation Modes}
\format{
Integer constants.

An object of class \code{integer} of length 1.

An object of class \code{integer} of length 1.
}
\usage{
NONE

PIPE

BUFFLE
}
\description{
These constants specify the mode used for calculating radiation impedance
in \code{calcimp(..., rad_calc = 'NONE/BUFFLE/PIPE')}.
}
\details{
The available modes are:
\itemize{
\item \code{NONE}: Do not compute radiation characteristics; simple open end.
\item \code{PIPE}: Radiation characteristics of a pipe without a flange (default).
\item \code{BUFFLE}: Radiation characteristics of a pipe with an infinite flange.
}
}
\keyword{datasets}
