% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcimp.R
\name{struve1}
\alias{struve1}
\title{Struve function}
\usage{
struve1(x)
}
\arguments{
\item{x}{Numeric value at which to evaluate struve(1, x)}
}
\value{
Numeric value of struve(1, x)
}
\description{
Just to test struve(1,x) works as expected.
Sometimes due to symbol collision of gamma function in cephes and libm,
struve(1,x) returns wrong result.
}
\examples{
struve1(0.059978)  # Should return approximately 0.000763201
}
