% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcimp.R
\docType{data}
\name{NONE}
\alias{NONE}
\alias{PIPE}
\alias{BUFFLE}
\title{Radiation Impedance Calculation Modes}
\usage{
NONE

PIPE

BUFFLE
}
\description{
Constants for specifying radiation impedance calculation mode in calcimp()
}
\keyword{datasets}
